CRAS: dev_stream - Mix proper format.
Mix the format that matches the device.  This adds a format parameter
to dev_stream_mix so the correct format can be used when mixing.

BUG=chromium:219948
TEST=play through USB DAC that supports 32 bit samples only.

Reviewed-on: https://chromium-review.googlesource.com/232085
Commit-Queue: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>

+++ cras/src/server/cras_rstream.c
@@ -313,7 +313,7 @@ float cras_rstream_get_volume_scaler(struct cras_rstream *rstream)
 	return cras_shm_get_volume_scaler(shm);
 }
 
-int16_t *cras_rstream_get_readable_frames(struct cras_rstream *rstream,
+uint8_t *cras_rstream_get_readable_frames(struct cras_rstream *rstream,
 					  unsigned int offset,
 					  size_t *frames)
 {
